backend/requirements.txt
Flask
newspaper3k
python-dotenv
langchain
openai
pytest
langchain-openai
"lxml[html_clean]"